add batch insert result analisys

+++ db/src/main/resources/db/migrations/v-0.2/tables/swift_mt.sql
@@ -1,5 +1,5 @@
 create table test.swift_mt (
-	id int primary key, 
+	id bigint primary key, 
 	operdate date,
 	body text
 );
