+++ producer/src/main/java/com/mycompany/app/helpers/SwiftMTHelper.java
@@ -19,11 +19,11 @@ public class SwiftMTHelper {
         .append(":71A:SHA")
         .append("-}");
 
-    public static String createMT103(int message_id) {
+    public static String createMT103(long message_id) {
         return mt103.toString().replace("REFERENCE12345", String.valueOf(message_id));
     }
 
-    public static int parseMessageId(String mt) {
+    public static long parseMessageId(String mt) {
         int idx1 = mt.indexOf(":20:") + 4;
         int idx2 = mt.indexOf(":23B:");
         return Integer.parseInt(mt.substring(idx1, idx2));
